MAVENPLUGIN-32 : fix compilation failure with maven >= 3.2.2
maven 3.2.2 uses by default maven-plugin-plugin >= 3.0.0 where
the annotation processors was added. The simple fix is to force
using maven-plugin-plugin 2.9, which is what happens when we use
maven-3.0.5

A longer fix is to finish c16be6a127b which added annotations
but we are not using them. We could configure maven-plugin-plugin
to use only one extractor (javadoc or annotations) but instead of
duplicating the information we should migrate everything to the
annotations.

Note: between maven 3.0.5 and 3.5.0 the default lifecycle binding
for the descriptor goal has moved from generate-resources to
process-classes (to allow using the annotation extractor) but this
shouldn't impact us.